refactor(conversation): extract the conversation serializer from the adapter
`agent-server-adapter.ts` is 1,278 lines mixing serialization, request
construction, agent-settings assembly, secret handling and auth validation.
#15533 sequences the decomposition: compatibility facade first, serializer
first because it is the cleanest seam. This is that slice.

Move `DirectConversationInfo`, `toConversationUrl`,
`getDefaultConversationTitle`, `toAppConversation` and `toConversationPage`
into `conversation-service/conversation-serializer.ts` verbatim — every moved
declaration is byte-identical to its original.

The wire tag keys move to `conversation-service/conversation-tag-keys.ts`
because both sides need them as runtime values: the adapter builds
`RESERVED_CONVERSATION_TAG_KEYS` from them and stamps `ACP_SERVER_TAG_KEY`
onto the start payload, while the serializer reads that same key back when
deriving `acp_server`. A shared module is the right home for a constant with
two real consumers, and it keeps either side from importing the other.

The adapter re-exports both modules so existing importers keep working, and
four callers move onto the narrow modules: the conversation service, the fork
hook, the conversation-panel list helpers, and the conversation mock.

Adds a serializer suite covering the seam plus direct tests for
`toConversationUrl` and `toConversationPage`, which had only indirect coverage
through the service layer.

No behaviour change; the conversation-start payload is unchanged.
